Sprint1V0: FIX conversione da int a float FW

+++ projects/unibo.serviceaccessGUISim/src/main/java/unibo/springSAGSim/connection/SagConnection.java
@@ -25,7 +25,7 @@ public class SagConnection extends ConnectionUtils{
         return connectActorUsingCoap(actorCtxHost, actorCtxPort, actorCtx, actorName);
     }
 
-    public String sendStorageRequest(Interaction conn, int fw){
+    public String sendStorageRequest(Interaction conn, float fw){
         String answer = "";
         try {
             String msg = ""+ CommUtils.buildRequest("ServiceAccessGUISim",
